requirements.txt
yfinance
pandas
numpy
matplotlib
scipy
jupyter
transformers
torch
python-dotenv
requests
alpaca-py
